@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Linker::Application.config.secret_token = '233412d2f23d058675af492144faf9c44cd9b1062ee833f14463c4ce75dacaeb5223c811f5819fdf8bc44faa6e71091476856b6de38bb2b5f32a2e1133b38ead'
+Linker::Application.config.secret_token = '524f146b1d2fd2739c37e31e8bbe899e5a87e63a77637684310ece92b2466fd67da8486fc50012ef1940b9bf13329b0439ffd38390accb3870e3bccc1345303a'
